move hyperbahn-client out of hyperbahn folder
